feat(markdown): allow disabling table `tabindex` attribute
Moves the inline table_open rule into its own plugin file and adds a
`markdown.tableTabIndex` option (default true) to disable it.
